feat(SPV-1425): add validation to paymail
