+++ actions/v2/admin/errors/errors.go
@@ -11,6 +11,9 @@ var ErrCreatingUser = models.SPVError{Message: "error creating user", StatusCode
 // ErrInvalidPaymail is returned when the paymail is invalid
 var ErrInvalidPaymail = models.SPVError{Message: "invalid paymail", StatusCode: 400, Code: "error-user-invalid-paymail"}
 
+// ErrInvalidAvatarURL is returned when the avatar url is invalid
+var ErrInvalidAvatarURL = models.SPVError{Message: "invalid avatar url", StatusCode: 400, Code: "error-user-invalid-avatar-url"}
+
 // ErrAddingPaymail is returned when the paymail addition fails
 var ErrAddingPaymail = models.SPVError{Message: "error adding paymail", StatusCode: 500, Code: "error-user-adding-paymail"}
