+++ engine/v2/users/usersmodels/user.go
@@ -5,12 +5,13 @@ import (
 	"time"
 
 	primitives "github.com/bitcoin-sv/go-sdk/primitives/ec"
+	"github.com/bitcoin-sv/spv-wallet/engine/v2/paymails/paymailsmodels"
 )
 
 // NewUser represents data for creating a new user
 type NewUser struct {
 	PublicKey string
-	Paymail   *NewPaymail
+	Paymail   *paymailsmodels.NewPaymail
 }
 
 // User is a domain model for existing users
@@ -20,7 +21,7 @@ type User struct {
 	UpdatedAt time.Time
 
 	PublicKey string
-	Paymails  []*Paymail
+	Paymails  []*paymailsmodels.Paymail
 }
 
 // PubKeyObj returns the go-sdk primitives.PublicKey object from the user's PubKey string
